backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic[email]==2.10.6
pydantic-core==2.27.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
PyPDF2==3.0.1
httpx==0.27.2
python-dotenv==1.0.1
resend==2.10.0
bleach==6.2.0
boto3==1.35.99
apscheduler==3.10.4
openai==1.58.1
slowapi==0.1.9
